feat: First simulation without learning
The function choose_action in agent.py has been
updated to be able to perform random actions at each
 time step, without learning and without using its
environment to determine which action to perform.
The flags in the function run, in the same python file,
have been updated to go through the simulation
quickly and generate the log file sim_no-learning.csv.